Fixed goods received

+++ payment/views.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from json import JSONDecodeError
 import traceback
 from django.shortcuts import render
@@ -154,7 +155,7 @@ class GoodsReceived(APIView):
         if transfer_code and status:
             Transaction.objects.create(
                 bank_details=detail,
-                amount=amount - (amount * 0.01),
+                amount=Decimal(amount) - Decimal((amount * 0.01)),
                 status=status,
                 transfer_code=transfer_code
             )
